refactor: better organization of the waPC runtime
Split the waPC runtime into smaller and dedicated files.
This now arranged like all the other runtimes we have.
