+++ src/runtimes/wapc/mod.rs
@@ -1,0 +1,10 @@
+mod callback;
+
+mod stack;
+pub(crate) use stack::WapcStack;
+
+mod runtime;
+pub(crate) use runtime::Runtime;
+
+mod mapping;
+pub(crate) use mapping::WAPC_POLICY_MAPPING;
